Remove excess modules, consolidate everything, get closer to old project to prepare for a PR
